Add: User deletion confirmation email

+++ controllers/userController.js
@@ -12,7 +12,7 @@ const {
   sendUnsubscribeEmail,
 } = require('./mailerController');
 
-// TODO: Poner o quitar favoritos
+// COMPLETE: Poner o quitar favoritos
 // TODO: Actualización de datos de usuario
 
 class UserController {
@@ -241,12 +241,14 @@ class UserController {
    */
   async deleteUser(req, res, next) {
     try {
-      // const user = await User.findOne({ _id: req.userId });
+      const { email } = await User.findOne({ _id: req.userId });
+
+      if (!email) return next(createError(404, 'User not found!'));
 
       await User.deleteOne({ _id: req.userId });
 
-      //TODO: Enviar email de confirmación de baja
-      // await sendUnsubscribeEmail(user.email);
+      //COMPLETE: Enviar email de confirmación de baja
+      await sendUnsubscribeEmail({ toUser: email });
 
       res.status(204).json({
         status: 'success',
